Move function to utility file
